requirements.txt
fastapi
uvicorn
python-multipart
openai
numpy
python-dotenv
